pyproject.toml
[tool.poetry]
name = "mailingpy"
version = "0.1.0"
description = ""
authors = ["Николай Некрасов <nnnekrasov@edu.hse.ru>"]
readme = "README.md"
packages = [{include = "mailingpy", from = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
sqlalchemy = "^2.0.17"
alembic = "^1.11.1"
psycopg2 = "^2.9.6"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
